libs/common: Refactor tempfile code into its own file
